--- src/components/popups-etc/tutorial.module.css
@@ -1,117 +1,0 @@
-.container {
-	z-index: 99;
-	position: fixed;
-	bottom: 0;
-	right: 10px;
-	display: flex;
-	flex-direction: column;
-	align-items: flex-end;
-	width: 500px;
-}
-
-.tabs {
-	display: flex;
-	flex-direction: row;
-}
-
-.tab {
-	background: var(--accent);
-	width: fit-content;
-	color: #ffffff;
-	padding: 6px 10px;
-	display: flex;
-	flex-direction: row;
-	gap: 4px;
-	align-items: center;
-	user-select: none;
-	cursor: var(--pointer-interactive);
-	z-index: 1;
-	border: solid var(--accent-dark);
-	border-width: 0;
-	border-right-width: 1px;
-	border-left-width: 1px;
-}
-
-.tab:first-of-type {
-	border-left-width: 0;
-}
-
-.tab:last-of-type {
-	border-right-width: 0;
-}
-
-.tab.selected {
-	background: var(--accent-dark);
-}
-
-.tab:hover {
-	background: var(--accent-dark);
-}
-
-.content {
-	background: var(--bg-surface);
-	border: 2px solid var(--accent);
-	border-bottom: none;
-	padding: 10px;
-	width: 100%;
-	overflow: auto;
-	height: 100%;
-	max-height: 500px;
-	box-shadow: 0 0 20px 0 #00000038;
-}
-
-.content strong {
-	color: #000000;
-}
-
-.content h1 {
-	text-align: center;
-	margin: 0;
-	margin-top: 20px;
-	margin-bottom: 30px;
-	color: #868e96;
-	font-size: 2.8em;
-	line-height: 1;
-	font-weight: normal;
-}
-
-.content h2 {
-	color: var(--accent-dark);
-	margin: 0;
-	font-size: 1.75em;
-	margin-top: 40px;
-	margin-bottom: 20px;
-}
-
-.content h3 {
-	margin: 0;
-	margin-top: 30px;
-	font-size: 1.125em;
-}
-
-.content p, .content ul, .content pre {
-	color: #495057;
-	margin: 16px 0;
-	line-height: var(--line-height-copy);
-}
-
-.content code, .content pre {
-	font-family: var(--font-code);
-	font-size: 0.9rem;
-	background: #e9ecef;
-	color: #343a40;
-}
-
-.content code {
-	padding: 0 4px;
-}
-
-.content pre code {
-	display: block;
-	padding: 12px 16px;
-	overflow: auto;
-}
-
-.content a:hover {
-	text-decoration: underline;
-}
